requirements.txt
flake8>=5.0.4, <5.0.5
Django>=4.1, <4.1.1
djangorestframework>=3.14.0, <3.14.1
psycopg2>=2.9.5, <3.0
Pillow>=9.3.0, <9.4
pyheif-pillow-opener>=0.1.0, <0.1.1
djangorestframework-simplejwt>=5.2.2, <5.2.3